Redo (and Redo's undo) now use the compressed map function
