+++ src/components/edit-tool.vue
@@ -1,5 +1,6 @@
 <script>
     import ToolbarButton from './toolbar-button.vue';
+    import UndoCalc from '../mixins/undo-calc.vue';
 
     export default {
         data() {
@@ -9,7 +10,7 @@
             }
         },
         extends: ToolbarButton,
-        mixins: [],
+        mixins: [UndoCalc],
         mounted() {
             this.$root.$on('deselect', this.deselect);
         },
@@ -69,72 +70,12 @@
                         undoMap[val].push(offset);
                     }
                 }
-                window.undoMap = undoMap;
                 // Try to compress the undo map using ranges where they are found
-                let compressedUndoMap = {};
-                let keyIndex, key, calcOffset, lowerOffset = null, lastOffset = null, index, array, compressedArray;
-                let undoMapKeys = Object.keys(undoMap);
-                for (keyIndex in undoMapKeys) {
-                    key = undoMapKeys[keyIndex];
-                    compressedArray = [];
-                    compressedUndoMap[key] = compressedArray;
-                    array = undoMap[key];
-                    // Reset our state
-                    lowerOffset = null;
-                    lastOffset = null;
-                    calcOffset = null;
-                    for (index = 0; index < array.length; index++) {
-                        offset = array[index];
-                        if (offset === calcOffset) { // A range is found
-                            if (lowerOffset === null) { // If this is the start of the range ...
-                                lowerOffset = lastOffset; // ... then mark the start
-                            }
-                        }
-                        else { // Not the continuation of a range
-                            if (lowerOffset !== null) { // If we just finished iterating through a range ...
-                                compressedArray.push([lowerOffset, lastOffset]); // ... then add the range
-                                lowerOffset = null;
-                            }
-                            else if (lastOffset !== null) {
-                                compressedArray.push(lastOffset); // Otherwise just add the last offset
-                            }
-                        }
-                        lastOffset = offset;
-                        calcOffset = offset + 4;
-                    }
-                    // Now we have to check the very last point
-                    if (lowerOffset !== null) { // If we were already in a range ...
-                        compressedArray.push([lowerOffset, offset]); // ... close it up
-                    }
-                    else {
-                        compressedArray.push(offset); // Otherwise just add the single value
-                    }
-                };
-                window.compressedUndoMap = compressedUndoMap;
+                let compressedUndoMap = this.compressUndoMap(undoMap);
                 this.$root.paint.state.saveUndoMap(compressedUndoMap);
                 canvas = this.$root.paint.canvas.undoElement;
                 this.$root.paint.canvas.drawContext.drawImage(canvas, 0, 0);
             }
         }
-    };
-
-    window.updateUndoMap = () => {
-        Object.keys(compressedUndoMap).forEach(key => {
-            let array = compressedUndoMap[key]; 
-            rebuiltUndoMap[key] = [];
-            array.forEach(elem => {
-            if (Array.isArray(elem)) {
-                let start = elem[0];
-                let end = elem[1];
-                let index;
-                for (index = start; index <= end; index += 4) {
-                rebuiltUndoMap[key].push(index);
-                }
-            }
-            else {
-                rebuiltUndoMap[key].push(elem);
-            }
-            });
-        });
     };
 </script>
